add a backend class that handles db and submission triggering
