+++ src/discord-cluster-manager/bot.py
@@ -7,6 +7,7 @@ import discord
 import env
 import uvicorn
 from api.main import app, init_api
+from backend import KernelBackend
 from cogs.admin_cog import AdminCog
 from cogs.leaderboard_cog import LeaderboardCog
 from cogs.misc_cog import BotManagerCog
@@ -15,24 +16,16 @@ from cogs.verify_run_cog import VerifyRunCog
 from discord import app_commands
 from discord.ext import commands
 from env import (
-    DATABASE_URL,
-    DISABLE_SSL,
     DISCORD_CLUSTER_STAGING_ID,
     DISCORD_DEBUG_CLUSTER_STAGING_ID,
     DISCORD_DEBUG_TOKEN,
     DISCORD_TOKEN,
-    POSTGRES_DATABASE,
-    POSTGRES_HOST,
-    POSTGRES_PASSWORD,
-    POSTGRES_PORT,
-    POSTGRES_USER,
     init_environment,
 )
 from launchers import GitHubLauncher, ModalLauncher
-from leaderboard_db import LeaderboardDB
 from utils import setup_logging
 
-logger = setup_logging()
+logger = setup_logging(__name__)
 
 
 class ClusterBot(commands.Bot):
@@ -58,35 +51,21 @@ class ClusterBot(commands.Bot):
         self.tree.add_command(self.admin_group)
         self.tree.add_command(self.leaderboard_group)
 
-        self.leaderboard_db = LeaderboardDB(
-            POSTGRES_HOST,
-            POSTGRES_DATABASE,
-            POSTGRES_USER,
-            POSTGRES_PASSWORD,
-            POSTGRES_PORT,
-            url=DATABASE_URL,
-            ssl_mode="require" if not DISABLE_SSL else "disable",
+        self.backend = KernelBackend(debug_mode=debug_mode)
+        self.backend.register_launcher(ModalLauncher(consts.MODAL_CUDA_INCLUDE_DIRS))
+        self.backend.register_launcher(
+            GitHubLauncher(env.GITHUB_REPO, env.GITHUB_TOKEN, env.GITHUB_WORKFLOW_BRANCH)
         )
 
-        try:
-            if not self.leaderboard_db.connect():
-                logger.error("Could not connect to database, shutting down")
-                exit(1)
-        finally:
-            self.leaderboard_db.disconnect()
-
-        self.accepts_jobs = True
+    @property
+    def leaderboard_db(self):
+        return self.backend.db
 
     async def setup_hook(self):
         logger.info(f"Syncing commands for staging guild {DISCORD_CLUSTER_STAGING_ID}")
         try:
             # Load cogs
-            submit_cog = SubmitCog(self)
-            submit_cog.register_launcher(ModalLauncher(consts.MODAL_CUDA_INCLUDE_DIRS))
-            submit_cog.register_launcher(
-                GitHubLauncher(env.GITHUB_REPO, env.GITHUB_TOKEN, env.GITHUB_WORKFLOW_BRANCH)
-            )
-            await self.add_cog(submit_cog)
+            await self.add_cog(SubmitCog(self))
             await self.add_cog(BotManagerCog(self))
             await self.add_cog(LeaderboardCog(self))
             await self.add_cog(VerifyRunCog(self))
@@ -251,7 +230,7 @@ async def start_bot_and_api(debug_mode: bool):
 
     bot_instance = ClusterBot(debug_mode=debug_mode)
 
-    init_api(bot_instance)
+    init_api(bot_instance.backend)
 
     config = uvicorn.Config(
         app,
